+++ .pm/tasks/task_5c651f038f1b48e78460ad7f95f6e187.execution.md
@@ -1,0 +1,19 @@
+# task_5c651f038f1b48e78460ad7f95f6e187 Execution Log
+
+- task_uid: task_5c651f038f1b48e78460ad7f95f6e187
+- title: burn down oasis7 node types controller binding hotspot
+- owner_role: runtime_engineer
+- worktree_hint: /home/scc/worktrees/oasis7-engineering-rust-node-types-controller-binding-burn-down
+
+<!-- Append entries using:
+Example:
+  ## YYYY-MM-DD HH:MM:SS CST / role_name
+  - 完成内容: ...
+  - 遗留事项: ...
+-->
+
+## 2026-04-17 21:53:46 CST / runtime_engineer
+- 完成内容: 将 `crates/oasis7_node/src/types.rs` 内的 main-token controller binding 常量、类型、默认值、builder 与校验 helper 抽到新文件 `crates/oasis7_node/src/types/main_token_controller_binding.rs`，并在 `types.rs` 中改为模块重导出；根文件已从 1277 行降到 1069 行，未引入新的 structural slicing / allowlist 债，同时已从 `doc/.governance/rust-oversized-file-baseline.tsv` 退休对应冻结条目。
+- 完成内容: 已执行 `env -u RUSTC_WRAPPER cargo fmt --all`、`env -u RUSTC_WRAPPER cargo check -p oasis7_node`、`cargo test -p oasis7_node types::tests:: -- --nocapture`、`cargo test -p oasis7_node submit_consensus_action_payload_accepts_signed_main_token_genesis_action -- --nocapture` 与 `cargo test -p oasis7_node submit_consensus_action_payload_accepts_signed_main_token_treasury_action -- --nocapture`，均通过，证明 controller binding 导出与授权路径未回归。
+- 完成内容: `cargo test -p oasis7_node --lib -- --nocapture` 当前环境仍有 3 个非定向失败签名：`runtime_network_replication_respects_topic_isolation`（目录残留断言）、`runtime_gossip_replication_persists_guard_across_restart`（UDP 端口占用）、`runtime_fetch_handlers_reject_unsigned_fetch_request_in_signed_mode`（`NetworkProtocolUnavailable`）；本轮未触碰对应 replication/network 行为，先按现状记录，不在本 task 扩大修复范围。
+- 遗留事项: 下一批 Rust 1200 行治理继续优先处理 `crates/oasis7/src/viewer/runtime_live/llm_sidecar.rs`、`crates/oasis7/src/bin/oasis7_web_launcher/control_plane.rs`、`crates/oasis7_client_launcher/src/launcher_core.rs`、`crates/oasis7_node/src/replication.rs` 与剩余高体量测试文件；`oasis7_game_launcher.rs` 当前实测 916 行，不再列为超限热点。
