+++ .pm/tasks/task_5c651f038f1b48e78460ad7f95f6e187.yaml
@@ -1,0 +1,26 @@
+task_uid: task_5c651f038f1b48e78460ad7f95f6e187
+title: "burn down oasis7 node types controller binding hotspot"
+owner_role: runtime_engineer
+worktree_hint: /home/scc/worktrees/oasis7-engineering-rust-node-types-controller-binding-burn-down
+execution_log_path: .pm/tasks/task_5c651f038f1b48e78460ad7f95f6e187.execution.md
+status: done
+priority: P2
+source_signal: null
+source_refs:
+  - doc/engineering/project.md
+  - doc/engineering/rust-governance/rust-1200-line-root-cause-governance-2026-03-29.project.md
+doc_refs:
+  - doc/engineering/rust-governance/rust-1200-line-root-cause-governance-2026-03-29.prd.md
+  - doc/engineering/rust-governance/rust-1200-line-root-cause-governance-2026-03-29.design.md
+related_prd:
+  - PRD-ENGINEERING-R1200-002
+  - PRD-ENGINEERING-R1200-004
+  - PRD-ENGINEERING-R1200-005
+acceptance:
+  - "reduce crates/oasis7_node/src/types.rs below the 1200-line limit without structural slicing debt"
+  - "preserve node controller binding validation behavior and public exports"
+  - "keep targeted oasis7_node validation green and structural slicing baseline unchanged"
+handoff_to: []
+updated_at: 2026-04-17T21:57:03+08:00
+last_started_at: 2026-04-17T21:49:20+08:00
+last_closed_at: 2026-04-17T21:57:03+08:00
